+++ Source/RTEngine/Sphere.hpp
@@ -8,18 +8,18 @@ namespace RealRT
     class Sphere : public Shape
     {
 		public:
-		    Sphere(const Material &mat, const Vector3D &center = {0.f, 0.f, 0.f}, double radius = 2.5);
+		    Sphere(const Material &mat, const Vector3D &center = {0.f, 0.f, 0.f}, float radius = 2.5f);
 		    ~Sphere(void) = default;
 
-		    double Intersect(const Ray &incident, bool &flipNormals) const;
+		    float Intersect(const Ray &incident, bool &flipNormals) const;
 		    Vector3D Normal(const Vector3D &pt) const;
 
-		    double Radius(void) const;
+		    float Radius(void) const;
 		    Vector3D Center(void) const;
 
 		private:
 		    Vector3D _Center;
-		    double _Radius;
+		    float _Radius;
     };
 
 }
